Move gfx1151 coverage into existing MXFP4 GEMM tests
